Update train_system.py
Added "speed_policy_reward_2" - Can't give speed that covers a distance greater than the one between two adjacent stations in one epoch.

Note: it may be less than an epoch so I generalized the condition. At this moment can't intergrate in "speed_rewarding" wrapper because full epoch size is not an argument of this function, we might need to save it as self.epoch and initialize it when calling simulation constructor in main - lets talk about it.